refactor: 태스크 추천 기능 분리
